adjust GearController to use interface, add DI to startup, controller tests work again
