+++ ClimbingGearBackend/Infrastructure/EFGearRepository.cs
