+++ ClimbingGearBackend/Interfaces/IGearRepository.cs
